requirements.txt
future-fstrings==1.2.0
db-sync-tool-kmi>=2.2.5
setuptools~=50.3.0